requirements.txt
accelerate==0.21.0
aiofiles==23.2.1
aiohttp==3.9.0
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
backoff==2.2.1
bcrypt==4.0.1
bitsandbytes==0.41.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.17
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.0
cycler==0.12.1
dataclasses-json==0.6.2
Deprecated==1.2.14
distro==1.8.0
einops==0.6.1
einops-exts==0.0.4
et-xmlfile==1.1.0
fastapi==0.104.1
ffmpy==0.3.1
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.44.3
frozenlist==1.4.0
fsspec==2023.10.0
google-auth==2.23.4
googleapis-common-protos==1.61.0
gradio==4.7.1
gradio_client==0.7.0
grpcio==1.59.3
h11==0.14.0
httpcore==0.17.3
httptools==0.6.1
httpx==0.24.0
huggingface-hub==0.19.4
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
kiwisolver==1.4.5
kubernetes==28.1.0
langchain==0.0.339
langsmith==0.0.66
linkify-it-py==2.0.2
-e git+https://github.com/haotian-liu/LLaVA.git@2ca20de1ca76d7d121be5a53f8a46c6bef47a9cb#egg=llava
markdown-it-py==2.2.0
markdown2==2.4.10
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.2
mdit-py-plugins==0.3.3
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.2
oauthlib==3.2.2
onnxruntime==1.16.3
openai==1.3.6
openpyxl==3.1.2
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
orjson==3.9.10
overrides==7.4.0
packaging==23.2
pandas==2.1.3
peft==0.4.0
Pillow==10.1.0
posthog==3.0.2
protobuf==4.25.1
psutil==5.9.6
pulsar-client==3.3.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.2
pydantic_core==2.14.5
pydub==0.25.1
Pygments==2.17.1
pyparsing==3.1.1
PyPika==0.48.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.0
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.13.0
rsa==4.9
safetensors==0.4.0
scikit-learn==1.2.2
scipy==1.11.4
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
shellingham==1.5.4
shortuuid==1.0.11
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
svgwrite==1.4.3
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.5.1
timm==0.6.13
tokenizers==0.13.3
tomlkit==0.12.0
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.31.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
uc-micro-py==1.0.2
urllib3==1.26.18
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
wavedrom==2.0.3.post3
websocket-client==1.6.4
websockets==11.0.3
wrapt==1.16.0
yarl==1.9.3
zipp==3.17.0
